[Refactor]: Change cae mae pretrain vit to cae mae vit
